+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/cargo/CargoInputNode.java
@@ -1,5 +1,8 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.cargo;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.inventory.ItemStack;
@@ -14,9 +17,12 @@ import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 
 public class CargoInputNode extends AbstractFilterNode {
 
-    private static final int[] BORDER = { 0, 1, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 16, 17, 18, 22, 23, 26, 27, 31, 32, 33, 34, 35, 36, 40, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53 };
-    private static final String ROUND_ROBIN_MODE = "round-robin";
+    private static final int[] BORDER = { 0, 1, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 17, 18, 22, 23, 26, 27, 31, 32, 33, 34, 35, 36, 40, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53 };
 
+    private static final String ROUND_ROBIN_MODE = "round-robin";
+    private static final String SMART_FILL_MODE = "smart-fill";
+
+    @ParametersAreNonnullByDefault
     public CargoInputNode(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe, ItemStack recipeOutput) {
         super(category, item, recipeType, recipe, recipeOutput);
     }
@@ -29,7 +35,9 @@ public class CargoInputNode extends AbstractFilterNode {
     @Override
     protected void onPlace(BlockPlaceEvent e) {
         super.onPlace(e);
+
         BlockStorage.addBlockInfo(e.getBlock(), ROUND_ROBIN_MODE, String.valueOf(false));
+        BlockStorage.addBlockInfo(e.getBlock(), SMART_FILL_MODE, String.valueOf(false));
     }
 
     @Override
@@ -48,6 +56,23 @@ public class CargoInputNode extends AbstractFilterNode {
             menu.replaceExistingItem(24, new CustomItem(HeadTexture.ENERGY_REGULATOR.getAsItemStack(), "&7Round-Robin Mode: &2\u2714", "", "&e> Click to disable Round Robin Mode", "&e(Items will be equally distributed on the Channel)"));
             menu.addMenuClickHandler(24, (p, slot, item, action) -> {
                 BlockStorage.addBlockInfo(b, ROUND_ROBIN_MODE, String.valueOf(false));
+                updateBlockMenu(menu, b);
+                return false;
+            });
+        }
+
+        String smartFillNode = BlockStorage.getLocationInfo(b.getLocation(), SMART_FILL_MODE);
+        if (!BlockStorage.hasBlockInfo(b) || smartFillNode == null || smartFillNode.equals(String.valueOf(false))) {
+            menu.replaceExistingItem(16, new CustomItem(Material.WRITABLE_BOOK, "&7\"Smart-Filling\" Mode: &4\u2718", "", "&e> Click to enable \"Smart-Filling\" Mode", "", "&fIn this mode, the Cargo node will attempt", "&fto keep a constant amount of items", "&fin the inventory. This is not perfect", "&fand will still fill in empty slots that", "&fcome before a stack of a configured item."));
+            menu.addMenuClickHandler(16, (p, slot, item, action) -> {
+                BlockStorage.addBlockInfo(b, SMART_FILL_MODE, String.valueOf(true));
+                updateBlockMenu(menu, b);
+                return false;
+            });
+        } else {
+            menu.replaceExistingItem(16, new CustomItem(Material.WRITTEN_BOOK, "&7\"Smart-Filling\" Mode: &2\u2714", "", "&e> Click to disable \"Smart-Filling\" Mode", "", "&fIn this mode, the Cargo node will attempt", "&fto keep a constant amount of items", "&fin the inventory. This is not perfect", "&fand will still fill in empty slots that", "&fcome before a stack of a configured item."));
+            menu.addMenuClickHandler(16, (p, slot, item, action) -> {
+                BlockStorage.addBlockInfo(b, SMART_FILL_MODE, String.valueOf(false));
                 updateBlockMenu(menu, b);
                 return false;
             });