+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/auto_crafters/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains any electric machines related to automatic crafting tables.
- */
-package io.github.thebusybiscuit.slimefun4.implementation.items.electric.machines.auto_crafters;