+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/multiblocks/ArmorForge.java
@@ -52,7 +52,11 @@ public class ArmorForge extends AbstractCraftingTable {
                 }
             }
 
-            SlimefunPlugin.getLocalization().sendMessage(p, "machines.pattern-not-found", true);
+            if (inv.isEmpty()) {
+                SlimefunPlugin.getLocalization().sendMessage(p, "machines.inventory-empty", true);
+            } else {
+                SlimefunPlugin.getLocalization().sendMessage(p, "machines.pattern-not-found", true);
+            }
         }
     }